+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/LoginActivity.java
@@ -1,4 +1,4 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -7,6 +7,8 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import nl.rug.www.summerschool.R;
 
 /**
  * This class is an acitivty started when the application is opened.
