+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/announcement/AnnouncementPagerActivity.java
@@ -1,8 +1,7 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller.announcement;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -12,24 +11,28 @@ import android.support.v7.app.AppCompatActivity;
 
 import java.util.List;
 
+import nl.rug.www.summerschool.controller.ContentsLab;
+import nl.rug.www.summerschool.R;
+import nl.rug.www.summerschool.model.Announcement;
+
 /**
- * This class is an acitivty that allows the general information fragments on this to be slided by.
+ * This class is an acitivty that allows the announcement fragments on this to be slided by.
  *
  * @since 13/04/2017
  * @author Jeongkyun Oh
  */
 
-public class GeneralInfoPagerActivity extends AppCompatActivity {
+public class AnnouncementPagerActivity extends AppCompatActivity {
 
-    private static final String EXTRA_GENERAL_INFO_ID =
-            "nl.rug.www.summerschool.generalinfo_id";
+    private static final String EXTRA_ANNOUNCEMENT_ID =
+            "nl.rug.www.summerschool.announcement_id";
 
     private ViewPager mViewPager;
-    private List<GeneralInfo> mGeneralInfos;
+    private List<Announcement> mAnnouncements;
 
     public static Intent newIntent(Context packageContext, String content) {
-        Intent intent = new Intent(packageContext, GeneralInfoPagerActivity.class);
-        intent.putExtra(EXTRA_GENERAL_INFO_ID, content);
+        Intent intent = new Intent(packageContext, AnnouncementPagerActivity.class);
+        intent.putExtra(EXTRA_ANNOUNCEMENT_ID, content);
         return intent;
     }
 
@@ -38,27 +41,27 @@ public class GeneralInfoPagerActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_content_pager);
 
-        String generalInfoId = (String) getIntent().getSerializableExtra(EXTRA_GENERAL_INFO_ID);
+        String announcementId = (String) getIntent().getSerializableExtra(EXTRA_ANNOUNCEMENT_ID);
 
         mViewPager = (ViewPager) findViewById(R.id.content_view_pager);
 
-        mGeneralInfos = ContentsLab.get(this).getGeneralInfos();
+        mAnnouncements = ContentsLab.get(this).getAnnouncements();
         FragmentManager fragmentManager = getSupportFragmentManager();
         mViewPager.setAdapter(new FragmentStatePagerAdapter(fragmentManager) {
             @Override
             public Fragment getItem(int position) {
-                GeneralInfo generalInfo = mGeneralInfos.get(position);
-                return GeneralInfoFragment.newInstance(generalInfo.getId());
+                Announcement announcement = mAnnouncements.get(position);
+                return AnnouncementFragment.newInstance(announcement.getId());
             }
 
             @Override
             public int getCount() {
-                return mGeneralInfos.size();
+                return mAnnouncements.size();
             }
         });
 
-        for (int i = 0; i < mGeneralInfos.size(); i++) {
-            if (mGeneralInfos.get(i).getId().equals(generalInfoId)) {
+        for (int i = 0; i < mAnnouncements.size(); i++) {
+            if (mAnnouncements.get(i).getId().equals(announcementId)) {
                 mViewPager.setCurrentItem(i);
                 break;
             }
