+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/forum/ForumFragment.java
@@ -1,4 +1,4 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller.forum;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
