+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/lecturer/LecturerFragment.java
@@ -1,13 +1,16 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller.lecturer;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.text.Html;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import nl.rug.www.summerschool.controller.ContentsLab;
+import nl.rug.www.summerschool.R;
+import nl.rug.www.summerschool.model.Lecturer;
 
 /**
  * Lecturer fragment is to show the details of lecturer information
