+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/lecturer/LecturerPagerActivity.java
@@ -1,8 +1,7 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller.lecturer;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -12,24 +11,28 @@ import android.support.v7.app.AppCompatActivity;
 
 import java.util.List;
 
+import nl.rug.www.summerschool.controller.ContentsLab;
+import nl.rug.www.summerschool.R;
+import nl.rug.www.summerschool.model.Lecturer;
+
 /**
- * This class is an acitivty that allows the announcement fragments on this to be slided by.
+ * This class is an acitivty that allows the lecturer fragments on this to be slided by.
  *
  * @since 13/04/2017
  * @author Jeongkyun Oh
  */
 
-public class AnnouncementPagerActivity extends AppCompatActivity {
+public class LecturerPagerActivity extends AppCompatActivity {
 
-    private static final String EXTRA_ANNOUNCEMENT_ID =
-            "nl.rug.www.summerschool.announcement_id";
+    private static final String EXTRA_LECTURER_ID =
+            "nl.rug.www.summerschool.lecturer_id";
 
     private ViewPager mViewPager;
-    private List<Announcement> mAnnouncements;
+    private List<Lecturer> mLecturers;
 
     public static Intent newIntent(Context packageContext, String content) {
-        Intent intent = new Intent(packageContext, AnnouncementPagerActivity.class);
-        intent.putExtra(EXTRA_ANNOUNCEMENT_ID, content);
+        Intent intent = new Intent(packageContext, LecturerPagerActivity.class);
+        intent.putExtra(EXTRA_LECTURER_ID, content);
         return intent;
     }
 
@@ -38,27 +41,27 @@ public class AnnouncementPagerActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_content_pager);
 
-        String announcementId = (String) getIntent().getSerializableExtra(EXTRA_ANNOUNCEMENT_ID);
+        String lecturerId = (String) getIntent().getSerializableExtra(EXTRA_LECTURER_ID);
 
         mViewPager = (ViewPager) findViewById(R.id.content_view_pager);
 
-        mAnnouncements = ContentsLab.get(this).getAnnouncements();
+        mLecturers = ContentsLab.get(this).getLecturers();
         FragmentManager fragmentManager = getSupportFragmentManager();
         mViewPager.setAdapter(new FragmentStatePagerAdapter(fragmentManager) {
             @Override
             public Fragment getItem(int position) {
-                Announcement announcement = mAnnouncements.get(position);
-                return AnnouncementFragment.newInstance(announcement.getId());
+                Lecturer lecturer = mLecturers.get(position);
+                return LecturerFragment.newInstance(lecturer.getId());
             }
 
             @Override
             public int getCount() {
-                return mAnnouncements.size();
+                return mLecturers.size();
             }
         });
 
-        for (int i = 0; i < mAnnouncements.size(); i++) {
-            if (mAnnouncements.get(i).getId().equals(announcementId)) {
+        for (int i = 0; i < mLecturers.size(); i++) {
+            if (mLecturers.get(i).getId().equals(lecturerId)) {
                 mViewPager.setCurrentItem(i);
                 break;
             }
