+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/controller/myprofile/MyProfileFragment.java
@@ -1,22 +1,19 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.controller.myprofile;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.facebook.login.LoginManager;
 import com.facebook.login.widget.ProfilePictureView;
 
-import org.w3c.dom.Text;
-
-import java.lang.reflect.Array;
 import java.util.ArrayList;
+
+import nl.rug.www.summerschool.R;
 
 /**
  * This class is to show the data fetched from facebook or google+.
