+++ SummerSchool/app/src/main/java/nl/rug/www/summerschool/model/TimeTable.java
@@ -1,4 +1,6 @@
-package nl.rug.www.summerschool;
+package nl.rug.www.summerschool.model;
+
+import nl.rug.www.summerschool.model.Content;
 
 /**
  * This class is a model of time table.
