+++ SummerSchool/app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:background="@color/colorPrimary"
-                tools:context="nl.rug.www.summerschool.LoginActivity">
+                tools:context="nl.rug.www.summerschool.controller.LoginActivity">
 
     <ImageView
         android:id="@+id/mainlogo"
